Move identifiers to macro
