+++ libs/@blockprotocol/type-system/rust/src/schema/data_type/mod.rs
@@ -28,47 +28,14 @@ use core::{fmt, mem};
 use std::collections::{HashMap, HashSet};
 
 use error_stack::{Report, bail};
-#[cfg(feature = "postgres")]
-use postgres_types::{FromSql, ToSql};
 use serde::{Deserialize, Serialize};
 use serde_json::Value as JsonValue;
 use thiserror::Error;
-#[cfg(feature = "utoipa")]
-use utoipa::ToSchema;
-use uuid::Uuid;
 
-use crate::{schema::data_type::constraint::ValueConstraints, url::VersionedUrl};
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[cfg_attr(feature = "utoipa", derive(ToSchema))]
-#[cfg_attr(feature = "postgres", derive(FromSql, ToSql), postgres(transparent))]
-#[repr(transparent)]
-pub struct DataTypeId(Uuid);
-
-impl DataTypeId {
-    #[must_use]
-    pub const fn new(uuid: Uuid) -> Self {
-        Self(uuid)
-    }
-
-    #[must_use]
-    pub fn from_url(url: &VersionedUrl) -> Self {
-        Self(Uuid::new_v5(
-            &Uuid::NAMESPACE_URL,
-            url.to_string().as_bytes(),
-        ))
-    }
-
-    #[must_use]
-    pub const fn as_uuid(&self) -> &Uuid {
-        &self.0
-    }
-
-    #[must_use]
-    pub const fn into_uuid(self) -> Uuid {
-        self.0
-    }
-}
+use crate::{
+    schema::{DataTypeId, data_type::constraint::ValueConstraints},
+    url::VersionedUrl,
+};
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
 #[cfg_attr(target_arch = "wasm32", derive(tsify::Tsify))]
