+++ libs/@blockprotocol/type-system/rust/src/schema/mod.rs
@@ -13,12 +13,14 @@ mod array;
 mod object;
 mod one_of;
 
+mod identifier;
+
 pub use self::{
     array::{PropertyArraySchema, PropertyValueArray, ValueOrArray},
     data_type::{
         AnyOfConstraints, ArrayConstraints, ArraySchema, ArrayTypeTag, ArrayValidationError,
         BooleanTypeTag, ClosedDataType, ConstraintError, ConversionDefinition,
-        ConversionExpression, ConversionValue, Conversions, DataType, DataTypeEdge, DataTypeId,
+        ConversionExpression, ConversionValue, Conversions, DataType, DataTypeEdge,
         DataTypeReference, DataTypeResolveData, DataTypeValidator, InheritanceDepth,
         JsonSchemaValueType, NullTypeTag, NumberConstraints, NumberSchema, NumberTypeTag,
         NumberValidationError, ObjectTypeTag, OntologyTypeResolver, Operator, ResolvedDataType,
@@ -27,9 +29,10 @@ pub use self::{
         ValidateDataTypeError, ValueLabel, ValueSchemaMetadata, Variable,
     },
     entity_type::{
-        ClosedEntityType, ClosedEntityTypeSchemaData, EntityType, EntityTypeId,
-        EntityTypeReference, EntityTypeValidationError, EntityTypeValidator,
+        ClosedEntityType, ClosedEntityTypeSchemaData, EntityType, EntityTypeReference,
+        EntityTypeValidationError, EntityTypeValidator,
     },
+    identifier::{DataTypeId, EntityTypeId, PropertyTypeId},
     object::{
         ObjectSchemaValidationError, ObjectSchemaValidator, PropertyObjectSchema,
         PropertyValueObject,
